backend/requirements.txt
# Azure Functions runtime
azure-functions

# Messaging (Service Bus)
azure-servicebus

# Metadata/state (Cosmos DB)
azure-cosmos

# input/output files
azure-storage-blob
